made IQM framerate a float
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@10052 d7cf8633-e32d-0410-b094-e92efae38249
::stable-branch::merge=c7b1d8b1697106c4a3616e0c09aef5e85ca7b39a

+++ model_iqm.h
@@ -74,7 +74,7 @@ typedef struct iqmanim_s
 {
 	unsigned int name;
 	unsigned int first_frame, num_frames;
-	unsigned int framerate;
+	float framerate;
 	unsigned int flags;
 }
 iqmanim_t;
